move OGRGetGeosVersion to ogr_api
