+++ gdal/ogr/ogr_api.h
@@ -49,6 +49,8 @@
 
 CPL_C_START
 
+void CPL_DLL OGRGetGEOSVersion(int *pnMajor, int *pnMinor, int *pnPatch);
+
 /* -------------------------------------------------------------------- */
 /*      Geometry related functions (ogr_geometry.h)                     */
 /* -------------------------------------------------------------------- */
